Refactor the pseudo_rands madness
For Argon2i, we need to generate pseudo-random addresses to compute the reference block number. The original code would dynamically allocate a buffer for these addresses, which was inefficient (and it would silently fail if the allocation did not succeed!!!). The new code computes the addresses on the fly, thus avoiding the need to allocate a buffer each time.